+++ xsimlab/variable.py
@@ -132,8 +132,15 @@ def variable(dims=(), intent='in', group=None, default=attr.NOTHING,
                 'attrs': attrs or {},
                 'description': description}
 
+    if VarIntent(intent) == VarIntent.OUT:
+        _init = False
+        _repr = False
+    else:
+        _init = True
+        _repr = True
+
     return attr.attrib(metadata=metadata, default=default, validator=validator,
-                       init=False, cmp=False, repr=False)
+                       init=_init, cmp=False, repr=_repr, kw_only=True)
 
 
 def on_demand(dims=(), group=None, description='', attrs=None):
@@ -229,7 +236,15 @@ def foreign(other_process_cls, var_name, intent='in'):
                 'intent': VarIntent(intent),
                 'description': description}
 
-    return attr.attrib(metadata=metadata, init=False, cmp=False, repr=False)
+    if VarIntent(intent) == VarIntent.OUT:
+        _init = False
+        _repr = False
+    else:
+        _init = True
+        _repr = True
+
+    return attr.attrib(metadata=metadata, init=_init, cmp=False, repr=_repr,
+                       kw_only=True)
 
 
 def group(name):
@@ -260,4 +275,5 @@ def group(name):
                 'intent': VarIntent.IN,
                 'description': description}
 
-    return attr.attrib(metadata=metadata, init=False, cmp=False, repr=False)
+    return attr.attrib(metadata=metadata, init=True, cmp=False, repr=True,
+                       default=tuple(), kw_only=True)